[UNI-353] refactor : 사용하지 않는 코드 삭제 (#237)
* [UNI-353] refactor : 사용하지 않는 코드 삭제

* [UNI-353] refactor : 사용하지 않는 코드, 주석 삭제

* [UNI-353] refactor : 사용하지 않는 클래스 삭제

* [UNI-353] refactor : 패키지 컨밴션에 맞도록 이동

* [UNI-353] refactor : 패키지 컨밴션에 맞도록 이동

* [UNI-353] refactor : 사용하지 않는 상수 삭제

* [UNI-353] refactor : 건물 이름을 strip() 에서 replaceAll로 변경

* [UNI-353] refactor : 로컬캐시 관련 코드 삭제

* [UNI-353] refactor : 컨버터 관련 코드 삭제
