+++ uniro_backend/src/main/java/com/softeer5/uniro_backend/common/config/WebMvcConfig.java
@@ -33,27 +33,6 @@ public class WebMvcConfig implements WebMvcConfigurer {
 		this.jwtInterceptor = jwtInterceptor;
 	}
 
-	// @Override
-	// public void configureMessageConverters(List<HttpMessageConverter<?>> converters) {
-	// 	FastJsonHttpMessageConverter converter = new FastJsonHttpMessageConverter();
-	// 	//custom configuration...
-	// 	FastJsonConfig config = new FastJsonConfig();
-	// 	config.setDateFormat("yyyy-MM-dd HH:mm:ss");
-	// 	config.setReaderFeatures(JSONReader.Feature.FieldBased, JSONReader.Feature.SupportArrayToBean);
-	// 	config.setWriterFeatures(JSONWriter.Feature.WriteMapNullValue, JSONWriter.Feature.PrettyFormat);
-	// 	converter.setFastJsonConfig(config);
-	// 	converter.setDefaultCharset(StandardCharsets.UTF_8);
-	//
-	// 	converter.setSupportedMediaTypes(List.of(
-	// 		MediaType.APPLICATION_JSON, // application/json 지원
-	// 		new MediaType("application", "json", StandardCharsets.UTF_8),
-	// 		new MediaType("application", "openmetrics-text", StandardCharsets.UTF_8)
-	// 	));
-	//
-	// 	converters.add(0, converter);
-	// 	converters.forEach(c -> log.info("✔ {}", c.getClass().getName()));
-	// }
-
 	@Override
 	public void addInterceptors(InterceptorRegistry registry) {
 		// JWT 인터셉터는 더 먼저 실행되도록 우선순위 낮춤
