+++ uniro_backend/src/main/java/com/softeer5/uniro_backend/map/service/RouteCalculator.java
@@ -2,8 +2,8 @@ package com.softeer5.uniro_backend.map.service;
 
 import static com.softeer5.uniro_backend.common.constant.UniroConst.*;
 import static com.softeer5.uniro_backend.common.error.ErrorCode.*;
-import static com.softeer5.uniro_backend.map.entity.RoadExclusionPolicy.calculateCost;
-import static com.softeer5.uniro_backend.map.entity.RoadExclusionPolicy.isAvailableRoute;
+import static com.softeer5.uniro_backend.map.enums.RoadExclusionPolicy.calculateCost;
+import static com.softeer5.uniro_backend.map.enums.RoadExclusionPolicy.isAvailableRoute;
 
 import com.softeer5.uniro_backend.common.error.ErrorCode;
 import com.softeer5.uniro_backend.common.exception.custom.NodeException;
@@ -14,11 +14,8 @@ import com.softeer5.uniro_backend.map.dto.FastestRouteResultDTO;
 import com.softeer5.uniro_backend.map.dto.response.*;
 import com.softeer5.uniro_backend.map.entity.*;
 import com.softeer5.uniro_backend.map.dto.request.CreateRouteReqDTO;
-import com.softeer5.uniro_backend.map.enums.CautionFactor;
-import com.softeer5.uniro_backend.map.enums.DangerFactor;
-import com.softeer5.uniro_backend.map.enums.DirectionType;
+import com.softeer5.uniro_backend.map.enums.*;
 import com.softeer5.uniro_backend.map.entity.Route;
-import com.softeer5.uniro_backend.map.enums.HeightStatus;
 import lombok.AllArgsConstructor;
 
 import org.locationtech.jts.geom.Coordinate;
@@ -77,10 +74,6 @@ public class RouteCalculator {
             adjMap.computeIfAbsent(route.getNode2().getId(), k -> new ArrayList<>()).add(route);
 
         }
-
-//        List<NodeInfoResDTO> nodeInfos = nodeMap.entrySet().stream()
-//                .map(entry -> NodeInfoResDTO.of(entry.getKey(), entry.getValue().getX(), entry.getValue().getY()))
-//                .toList();
 
         List<NodeInfoResDTO> nodeInfos = new ArrayList<>();
 
@@ -201,9 +194,6 @@ public class RouteCalculator {
         PriorityQueue<CostToNextNode> pq = new PriorityQueue<>();
         pq.add(new CostToNextNode(0.0, startNode,0));
         costMap.put(startNode.getId(), 0.0);
-        // A* 알고리즘 중 출발점과 도착점의 해발고도를 크게 벗어나지 않도록 하기 위한 변수 설정
-        final double minHeight = Math.min(startNode.getHeight(),endNode.getHeight());
-        final double maxHeight = Math.max(startNode.getHeight(),endNode.getHeight());
 
         // 길찾기 알고리즘
         while(!pq.isEmpty()){
@@ -293,10 +283,10 @@ public class RouteCalculator {
 
             double heightDiff = firstNode.getHeight() - secondNode.getHeight();
             if(heightDiff > 0){
-                heightIncreaseWeight += Math.max(LIMIT_RANGE ,Math.exp(heightDiff) - 1);
+                heightIncreaseWeight += Math.min(LIMIT_RANGE ,Math.exp(heightDiff) - 1);
             }
             else{
-                heightDecreaseWeight += Math.max(LIMIT_RANGE, Math.exp(-heightDiff) - 1);
+                heightDecreaseWeight += Math.min(LIMIT_RANGE, Math.exp(-heightDiff) - 1);
             }
 
             routeInfoDTOS.add(RouteInfoResDTO.of(route, firstNode, secondNode));
@@ -549,8 +539,7 @@ public class RouteCalculator {
     }
 
 
-    // TODO: 모든 점이 아니라 request 값의 MBR 영역만 불러오면 좋을 것 같다.  <- 추가 검증 필요
-    // TODO: 캐시 사용 검토
+
     public List<Node> createValidRouteNodes(Long univId, Long startNodeId, Long endNodeId, List<CreateRouteReqDTO> requests, List<Route> routes) {
         LinkedList<Node> createdNodes = new LinkedList<>();
